Major(Publish.js): Removes Publish.js Creation
Callouts created by Admonitions do more than an adequate job for Obsidian Publish. Admonitions generating a publish.js will be depreciated at this time.

BREAKING CHANGE: This is a technically a breaking change because it removes a "feature". However, this feature does not work on the current version of Obsidian Publish.

Signed-off-by: Sigrunixia <Scholarlysigrun@icloud.com>